+++ crates/daemon/src/fapolicyd.rs
@@ -11,9 +11,12 @@
 use crate::error::Error;
 use std::fs::File;
 use std::io::Read;
-use std::path::Path;
-use std::thread;
+use std::path::{Path, PathBuf};
+use std::process::{Child, Command, Stdio};
+use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::Arc;
 use std::time::Duration;
+use std::{io, thread};
 
 pub const TRUST_LMDB_PATH: &str = "/var/lib/fapolicyd";
 pub const TRUST_LMDB_NAME: &str = "trust.db";
@@ -28,6 +31,117 @@ pub const FIFO_PIPE: &str = "/run/fapolicyd/fapolicyd.fifo";
 pub enum Version {
     Unknown,
     Release { major: u8, minor: u8, patch: u8 },
+}
+
+pub struct Daemon {
+    pub name: String,
+    alive: Arc<AtomicBool>,
+    term: Arc<AtomicBool>,
+}
+
+impl Daemon {
+    pub fn new(name: &str) -> Self {
+        Self {
+            name: name.to_string(),
+            alive: Default::default(),
+            term: Default::default(),
+        }
+    }
+
+    pub fn active(&self) -> bool {
+        self.alive.load(Ordering::Relaxed)
+    }
+
+    pub fn stop(&self) {
+        self.term.store(true, Ordering::Relaxed)
+    }
+
+    pub fn start(&self, events_log: Option<&PathBuf>) -> io::Result<()> {
+        let (mut cmd, _) = build(
+            "/usr/sbin/fapolicyd --debug --permissive --no-details",
+            events_log,
+        );
+        let alive: Arc<AtomicBool> = self.alive.clone();
+        let term: Arc<AtomicBool> = self.term.clone();
+
+        thread::spawn(move || {
+            let mut execd = Execd::new(cmd.spawn().unwrap());
+
+            // the process is now alive
+            alive.store(true, Ordering::Relaxed);
+
+            while let Ok(true) = execd.running() {
+                thread::sleep(Duration::from_secs(1));
+                if term.load(Ordering::Relaxed) {
+                    execd.kill().expect("kill daemon");
+                    break;
+                }
+            }
+
+            // we need to wait on the process to die, instead of just blocking
+            // this loop provides the ability to add a harder stop impl, abort
+            term.store(false, Ordering::Relaxed);
+            while let Ok(true) = execd.running() {
+                if term.load(Ordering::Relaxed) {
+                    execd.kill().expect("abort daemon");
+                    break;
+                }
+                thread::sleep(Duration::from_secs(1));
+            }
+
+            // no longer alive
+            alive.store(false, Ordering::Relaxed);
+        });
+
+        Ok(())
+    }
+}
+
+type CmdArgs = (Command, String);
+fn build(args: &str, out: Option<&PathBuf>) -> CmdArgs {
+    let opts: Vec<&str> = args.split(' ').collect();
+    let (target, opts) = opts.split_first().expect("invalid cmd string");
+
+    let mut cmd = Command::new(target);
+
+    if let Some(path) = out {
+        println!("writing stderr to {}", path.display());
+        let mut f = File::create(path).unwrap();
+        cmd.stderr(Stdio::from(f));
+    }
+
+    cmd.args(opts);
+    (cmd, args.to_string())
+}
+
+/// Internal struct used to inspect a running process
+struct Execd {
+    proc: Option<Child>,
+}
+
+impl Execd {
+    fn new(proc: Child) -> Execd {
+        Execd { proc: Some(proc) }
+    }
+
+    /// Get the process id (pid), None if inactive
+    fn pid(&self) -> Option<u32> {
+        self.proc.as_ref().map(|p| p.id())
+    }
+
+    /// Is process is still running?, never blocks
+    fn running(&mut self) -> io::Result<bool> {
+        match self.proc.as_mut().unwrap().try_wait() {
+            Ok(Some(_)) => Ok(false),
+            Ok(None) => Ok(true),
+            Err(e) => Err(e),
+        }
+    }
+
+    /// Cancel the process, without blocking
+    fn kill(&mut self) -> io::Result<()> {
+        self.proc.as_mut().unwrap().kill()
+    }
 }
 
 /// watch a fapolicyd log at the specified path for the
