--- crates/daemon/src/profiler.service
@@ -1,12 +1,0 @@
-[Unit]
-Description=File Access Profiling Daemon
-DefaultDependencies=no
-After=local-fs.target systemd-tmpfiles-setup.service
-
-[Install]
-WantedBy=multi-user.target
-
-[Service]
-Type=simple
-PIDFile=/run/fapolicyp.pid
-ExecStart=bash -c 'exec /usr/sbin/fapolicyd --debug --permissive --no-details 2> /dev/null'
